toggle toolbar using ctrl+tab and reorganized files, added comments and made class node button color available
